Modify: Classes have been reorganized according to the type of persistence.
