+++ src/test/java/Ticket/Repository/TicketRepositoryTXTTest.java
@@ -1,14 +1,12 @@
 package Ticket.Repository;
 
 import org.flowershop.domain.tickets.Ticket;
-import org.flowershop.domain.tickets.TicketDetail;
-import org.flowershop.repository.TicketRepositoryTXT;
+import org.flowershop.repository.repositoryTXT.TicketRepositoryTXT;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
-import java.util.Date;
 import java.util.List;
